Gemini-Multi-Purpose-App/Backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
